+++ packages/bridge-status-controller/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Error logs for invalid getTxStatus responses ([#5816](https://github.com/MetaMask/core/pull/5816))
+- Bridge status controller now sets the polling duration to 5 minutes ([#5683](https://github.com/MetaMask/core/pull/5683))
 
 ### Changed